Refactor Photon
