+++ Patriot/Hardware/PhotonError.swift
@@ -1,0 +1,19 @@
+//
+//  PhotonError.swift
+//  Patriot
+//
+//  Created by Ron Lisle on 5/30/18.
+//  Copyright © 2018 Rons iMac. All rights reserved.
+//
+
+import Foundation
+
+enum PhotonError : Error
+{
+    case DeviceVariable
+    case SupportedVariable
+    case ActivitiesVariable
+    case PublishVariable
+    case ReadVariable
+    case unknown
+}
